Add files via upload
An ALU is a common component of many computer hardware systems. At it’s core the ALU is a combination
of different combinational logic operations, referred to as functional units, which you can select using either
a Multiplexer or a series of Tri-state Buffers tied to the output. For this project, I designed an
ALU with 8 functional units and a multiplexer to select which functional unit’s output to pass to the output
of the ALU.